fixed errors, able to build the frikin gpu version now! need to test all .cu files next

+++ libs/psbmpc_lib_gpu/src/ownship.cu
@@ -368,7 +368,7 @@ __host__ __device__ void Ownship::predict_trajectory(
 		if (k == maneuver_times[man_count]){
 			u_m = offset_sequence[2 * man_count];
 			chi_m = offset_sequence[2 * man_count + 1]; 
-			if (man_count < maneuver_times.size() - 1) man_count += 1;
+			if (man_count < (int)maneuver_times.size() - 1) man_count += 1;
 		}  
 
 		update_guidance_references(u_d_p, chi_d_p, waypoints, xs, dt, guidance_method);
